+++ src/Componenets/App/App.module.scss
@@ -1,3 +1,5 @@
+@import './variables';
+
 .app {
   direction: rtl;
   display: flex;
